Add files via upload
Updated main.py and predict.py for parts with SPEED.
Added 3 json files that would be read in as dictionary for the statistical tree that SPEED algorithm would use to calculate probabilities.

In line 111 of the main.py file, there is the variable "input_seq" that would be the input string as context for prediction. Please assign the right variable that should be sent to it there.
In line 112 of the main.py file, there would be two outputs to the prediction, "prediction" and "out_prob". Please modify the codes with them so to fit the format for the graph output.

Here is how "prediction" and "out_prob" look like.
out_prob = [('A', 5.233526671520322e-05), ('B', 9.67739795036514e-05), ('C', 0.0), ('D', 0.0), ('E', 0), ('F', 0), ('G', 0), ('H', 0), ('I', 0), ('J', 0.0), ('K', 0.0007148726187303128), ('L', 0.0011443348794418346), ('a', 0.0), ('b', 0.0), ('c', 2.3258710201597944e-06), ('d', 0.0028911147157675656), ('e', 0), ('f', 0), ('g', 0), ('h', 0), ('i', 0), ('j', 0.0), ('k', 0.0001784900038468132), ('l', 0.0009702494542299289)]
prediction = ('d', 0.0028911147157675656)